Modify Demo

+++ DemoCode/DemoCode/Demo/News/HyNewsViewModel.m
@@ -17,7 +17,7 @@
     [self configRequestIsGet:YES url:^NSString * _Nonnull(id  _Nullable input, HyListViewRequestDataType type) {
         @strongify(self);
         return
-        [NSString stringWithFormat:@"http://i.play.163.com/user/article/list/%ld/%ld", self.getRequestDataPageNumber(type) * self.getPageSize, (long)self.getPageSize];
+        [NSString stringWithFormat:@"http://i.play.163.com/user/article/list/%ld/%ld", (long)self.getRequestDataPageNumber(type) * self.getPageSize, (long)self.getPageSize];
     } parameter:nil sectionDataHandler:nil cellDataHandler:^NSArray<id> * _Nonnull(id  _Nullable input, NSDictionary * _Nonnull sectionData, NSUInteger section, HyListViewRequestDataType type) {
         return @[sectionData[@"info"], HyNewsModel.class];
     }];
